service_a/requirements.txt
Django==5.2.6
grpcio==1.75.0
djangorestframework==3.16.1
